Add a thread for simulation update

+++ glfw/src/Simulation/Tasks/UpdateSceneAndPhysicsTask.hpp
@@ -42,7 +42,7 @@ namespace crimild {
 		virtual void stop( void ) override;
         
 	private:
-		float _dtAccumulator;
+		double _dtAccumulator;
 	};
 
 }
